+++ src/main/kotlin/com/sota/clone/copyopgg/controllers/SummonerController.kt
@@ -22,9 +22,9 @@ import java.util.*
 class SummonerController(
     @Autowired val summonerRepo: SummonerRepository,
     @Autowired val leagueRepo: LeagueRepository,
-    @Autowired val leagueSummonerRepo: LeagueSummonerRepository
+    @Autowired val leagueSummonerRepo: LeagueSummonerRepository,
+    val riotApiController: RiotApiController,
 ) {
-
     val logger: Logger = LoggerFactory.getLogger(SummonerController::class.java)
 
     @GetMapping("/search/auto-complete/{searchWord}")
@@ -49,7 +49,7 @@ class SummonerController(
         val result = summonerRepo.searchByName(searchWord)
         return result?.let {
             ResponseEntity.ok().body(it)
-        } ?: this.getSummonerViaRiotApi(searchWord)?.let {
+        } ?: this.riotApiController.getSummoner(searchWord)?.let {
             summonerRepo.insertSummoner(it)
             ResponseEntity.ok().body(
                 SummonerBriefInfo(
@@ -61,21 +61,6 @@ class SummonerController(
                 )
             )
         } ?: ResponseEntity.notFound().build()
-    }
-
-    fun getSummonerViaRiotApi(searchWord: String): SummonerDTO? {
-        val fromOrianna = Orianna.summonerNamed(searchWord).withRegion(Region.KOREA).get()
-        return fromOrianna.puuid?.let {
-            SummonerDTO(
-                accountId = fromOrianna.accountId,
-                puuid = fromOrianna.puuid,
-                id = fromOrianna.id,
-                name = fromOrianna.name,
-                summonerLevel = fromOrianna.level.toLong(),
-                profileIconId = fromOrianna.profileIcon.id,
-                revisionDate = fromOrianna.updated.millis
-            )
-        } ?: null
     }
 
     @GetMapping("/league/brief/{searchId}")
@@ -100,6 +85,12 @@ class SummonerController(
                 )
             } ?: ResponseEntity.notFound().build()
         } ?: ResponseEntity.notFound().build()
+    }
+
+    @GetMapping("/refresh/{summonerId}")
+    fun refresh(@PathVariable(name = "summonerId", required = true) summonerId: String): ResponseEntity<BooleanResponse> {
+        logger.info("Synchronize data of summoner has id $summonerId")
+        return ResponseEntity.notFound().build()
     }
 }
