+++ src/main/kotlin/com/sota/clone/copyopgg/models/GeneralResponse.kt
@@ -1,0 +1,6 @@
+package com.sota.clone.copyopgg.models
+
+data class BooleanResponse(
+    val id: String,
+    val result: Boolean
+)
